impede de deletar

+++ src/Services/Users/destroyUser.ts
@@ -1,4 +1,4 @@
-import { NotFoundError } from "../../Utils/Errors/CError";
+import { ConflictError, NotFoundError } from "../../Utils/Errors/CError";
 import { prisma } from "../../Utils/prisma";
 import { omitUser } from "./Utils/functions";
 
@@ -9,6 +9,14 @@ export const destroyUser = async (id: number) => {
       omit: omitUser,
     });
     if (!user) throw new NotFoundError("Usuário não encontrado.");
+    const secretary = await prisma.secretaria.findUnique({
+      where: { secretario_id: id },
+    });
+    if (!!secretary)
+      throw new ConflictError(
+        "Esse usuário esta vinculado a uma secretaria, não é possível deletá-lo."
+      );
+
     await prisma.usuarios.delete({ where: { id } });
     return "Usuário deletado com sucesso.";
   } catch (error) {
